extract stateTransitions.test from se.test
