+++ sonarts-core/src/se/stateTransitions.ts
