+++ sonarts-core/tests/se/stateTransitions.test.ts
@@ -1,0 +1,29 @@
+/*
+ * SonarTS
+ * Copyright (C) 2017-2017 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+import { run } from "../utils/seTestUtils";
+
+describe("Expressions", () => {
+  it("chains assignments", () => {
+    expect.assertions(1);
+    run(`let a; let b; a = b = 0; _inspect(a);`, (node, states, symbols) => {
+      expect(states[0].sv(symbols.get("a"))).toEqual({ type: "literal", value: "0" });
+    });
+  });
+});
